Fix top level directories
As it currently is, if this repo is added to a project as a git repo as a submodule or just downloaded manually, it sticks the Readme, Docs, and License files *outside* of the plugins folder.

To fix this, the Content, Resources, and Source folders, as well as the UPlugin file, should be moved out of the folder they are in, and moved up to top. That way the entire contents of this repo will properly populate into a plugin folder when added as a submodule.